Add an ERROR command that triggers immediate reboot
Sometimes a client is able to detect an error that it can't recover from
except by rebooting. In that case, it's better to use the WDT to reboot
rather than attempting a reboot yourself, so the watchdog timer is
properly handled.

Add a new command ERROR to the socket interface. By calling ERROR
instead of KICK, a client can force a reboot (if the -c option was
given) or waiting for a watchdog reset (if the WDT exits without
reboot).